+++ ZoomSDK/ZoomSDK.framework/Versions/A/Headers/ZoomSDKBuildType.h
@@ -1,0 +1,4 @@
+#pragma once 
+
+#define BUILD_FOR_SDK 
+